refactor: split CLI audit and digest helpers
